Добавлен проект Visual Studio

+++ sources/Panel/src/Display/Draw.cpp
@@ -1,14 +1,16 @@
+#include "defines.h"
 #include "stm32f4xx_hal.h"
 #include "Draw.h"
 #include "Display/Display.h"
 
-void SetPoint(int x, int y, uint8_t color)
+
+void SetPoint(int x, int y, uint8 color)
 {
 	uint8_t* point = Display::GetBuff() + 320 * y + x;
 	*point = color;
 }
 
-void Draw::Rectangle(int x, int y, int height, int width, int color)
+void Draw::Rectangle(int x, int y, int height, int width, uint8 color)
 {
 	for (int i = x; i < x + width ; i++)
 	{
@@ -21,8 +23,8 @@ void Draw::Rectangle(int x, int y, int height, int width, int color)
 
 void Draw::Clear()
 {
-	uint32_t h = 240;
-	uint32_t w = 320;
+	int h = 240;
+	int w = 320;
 	for (int x = 0; x < w ; x++)
 	{
 		for (int y = 0; y < h; y++)
