+++ sources/Panel/src/Display/Draw.h
@@ -1,8 +1,10 @@
 #pragma once
+#include "defines.h"
+
 
 namespace Draw
 {
-    void Rectangle(int, int, int, int, int);
+    void Rectangle(int, int, int, int, uint8);
     void Clear();
 		
 };
